Move some pins on USB2504 for easier connecting
The strapping and LED pins being on the right are very difficult to
neatly connect up.

SMSC's reference design does this well but that design does not use the
strapping resistors as it has an EEPROM.  Adding the strapping resistors
into that same area as their design's LEDs would make it a mess.
Putting these pins on the left where not much else is makes things
(hopefully) a bit neater even if sometimes these pins are outputs and
now they're on the left.